--- x-pack/plugin/slm/src/main/java/org/elasticsearch/xpack/ilm/IndexLifecycle.java
@@ -1,88 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-package org.elasticsearch.xpack.ilm;
-
-import org.elasticsearch.cluster.metadata.Metadata;
-import org.elasticsearch.xcontent.NamedXContentRegistry;
-import org.elasticsearch.xcontent.ParseField;
-import org.elasticsearch.xpack.core.ilm.AllocateAction;
-import org.elasticsearch.xpack.core.ilm.DeleteAction;
-import org.elasticsearch.xpack.core.ilm.DownsampleAction;
-import org.elasticsearch.xpack.core.ilm.ForceMergeAction;
-import org.elasticsearch.xpack.core.ilm.FreezeAction;
-import org.elasticsearch.xpack.core.ilm.IndexLifecycleMetadata;
-import org.elasticsearch.xpack.core.ilm.LifecycleAction;
-import org.elasticsearch.xpack.core.ilm.LifecycleOperationMetadata;
-import org.elasticsearch.xpack.core.ilm.LifecycleType;
-import org.elasticsearch.xpack.core.ilm.MigrateAction;
-import org.elasticsearch.xpack.core.ilm.ReadOnlyAction;
-import org.elasticsearch.xpack.core.ilm.RolloverAction;
-import org.elasticsearch.xpack.core.ilm.SearchableSnapshotAction;
-import org.elasticsearch.xpack.core.ilm.SetPriorityAction;
-import org.elasticsearch.xpack.core.ilm.ShrinkAction;
-import org.elasticsearch.xpack.core.ilm.TimeseriesLifecycleType;
-import org.elasticsearch.xpack.core.ilm.UnfollowAction;
-import org.elasticsearch.xpack.core.ilm.WaitForSnapshotAction;
-import org.elasticsearch.xpack.core.slm.SnapshotLifecycleMetadata;
-
-import java.util.Arrays;
-import java.util.List;
-
-public class IndexLifecycle {
-
-    public static final List<NamedXContentRegistry.Entry> NAMED_X_CONTENT_ENTRIES = xContentEntries();
-
-    private static List<NamedXContentRegistry.Entry> xContentEntries() {
-        return Arrays.asList(
-            // Custom Metadata
-            new NamedXContentRegistry.Entry(
-                Metadata.Custom.class,
-                new ParseField(IndexLifecycleMetadata.TYPE),
-                parser -> IndexLifecycleMetadata.PARSER.parse(parser, null)
-            ),
-            new NamedXContentRegistry.Entry(
-                Metadata.Custom.class,
-                new ParseField(SnapshotLifecycleMetadata.TYPE),
-                parser -> SnapshotLifecycleMetadata.PARSER.parse(parser, null)
-            ),
-            new NamedXContentRegistry.Entry(
-                Metadata.Custom.class,
-                new ParseField(LifecycleOperationMetadata.TYPE),
-                parser -> LifecycleOperationMetadata.PARSER.parse(parser, null)
-            ),
-            // Lifecycle Types
-            new NamedXContentRegistry.Entry(
-                LifecycleType.class,
-                new ParseField(TimeseriesLifecycleType.TYPE),
-                (p, c) -> TimeseriesLifecycleType.INSTANCE
-            ),
-            // Lifecycle Actions
-            new NamedXContentRegistry.Entry(LifecycleAction.class, new ParseField(AllocateAction.NAME), AllocateAction::parse),
-            new NamedXContentRegistry.Entry(LifecycleAction.class, new ParseField(ForceMergeAction.NAME), ForceMergeAction::parse),
-            new NamedXContentRegistry.Entry(LifecycleAction.class, new ParseField(ReadOnlyAction.NAME), ReadOnlyAction::parse),
-            new NamedXContentRegistry.Entry(LifecycleAction.class, new ParseField(RolloverAction.NAME), RolloverAction::parse),
-            new NamedXContentRegistry.Entry(LifecycleAction.class, new ParseField(ShrinkAction.NAME), ShrinkAction::parse),
-            new NamedXContentRegistry.Entry(LifecycleAction.class, new ParseField(DeleteAction.NAME), DeleteAction::parse),
-            new NamedXContentRegistry.Entry(LifecycleAction.class, new ParseField(FreezeAction.NAME), FreezeAction::parse),
-            new NamedXContentRegistry.Entry(LifecycleAction.class, new ParseField(SetPriorityAction.NAME), SetPriorityAction::parse),
-            new NamedXContentRegistry.Entry(LifecycleAction.class, new ParseField(UnfollowAction.NAME), UnfollowAction::parse),
-            new NamedXContentRegistry.Entry(
-                LifecycleAction.class,
-                new ParseField(WaitForSnapshotAction.NAME),
-                WaitForSnapshotAction::parse
-            ),
-            new NamedXContentRegistry.Entry(
-                LifecycleAction.class,
-                new ParseField(SearchableSnapshotAction.NAME),
-                SearchableSnapshotAction::parse
-            ),
-            new NamedXContentRegistry.Entry(LifecycleAction.class, new ParseField(MigrateAction.NAME), MigrateAction::parse),
-            // TSDB Downsampling
-            new NamedXContentRegistry.Entry(LifecycleAction.class, new ParseField(DownsampleAction.NAME), DownsampleAction::parse)
-        );
-    }
-}
